bot/requirements.txt
discord.py==1.4.1
requests==2.24.0
environs==8.0.0
Pillow==7.2.0
python-dotenv==0.14.0
firebase-admin==4.4.0